added adding user logic, and some small changes in DB. Also added UserController, all entity services and repositories.

+++ src/main/java/com/teamchallenge/chat/entities/City.java
@@ -10,7 +10,7 @@ import lombok.Setter;
 public class City {
     @Id
     @Column(name = "id_city")
-    private int idCity;
+    private long idCity;
 
     @Column(name = "city_name")
     private String cityName;
